connect user with hibernate
