+++ server/movie_review/.gitignore
@@ -4,6 +4,8 @@ target/
 !**/src/main/**/target/
 !**/src/test/**/target/
 
+application.yaml
+
 ### STS ###
 .apt_generated
 .classpath
