+++ server/movie_review/src/main/java/com/ex/movie_review/MovieReviewApplication.java
@@ -1,5 +1,8 @@
 package com.ex.movie_review;
 
+import com.ex.movie_review.models.User;
+import com.ex.movie_review.repositories.UserRepository;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
@@ -13,6 +16,7 @@ public class MovieReviewApplication {
     private String value;
 
     public static void main(String[] args) {
+
         SpringApplication.run(MovieReviewApplication.class, args);
     }
