+++ server/movie_review/src/main/java/com/ex/movie_review/repositories/UserRepository.java
@@ -1,0 +1,9 @@
+package com.ex.movie_review.repositories;
+
+import com.ex.movie_review.models.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Integer> {
+}
