+++ server/movie_review/src/main/resources/application.yaml
@@ -6,3 +6,9 @@ server:
 my:
   message:
     value: 'Hello world!'
+spring:
+  datasource:
+    url: jdbc:postgresql://project0.c97dt2ne08zr.us-west-1.rds.amazonaws.com/postgres
+    username: postgres
+    password: AU85#f&EVKVW5iT
+    driver-class-name: org.postgresql.Driver
